Refactor: Move CardTag enum into TagsComponent
Affected files have been updated to reflect the new location of `CardTag`.
