+++ app/src/main/java/com/sq/thed_ck_licker/ecs/components/TagsComponent.kt
@@ -1,9 +1,10 @@
 package com.sq.thed_ck_licker.ecs.components
 
 
-enum class CardTag { CARD, MERCHANT, CORRUPTED }
-
 data class TagsComponent(private val tags: List<CardTag> = emptyList()) {
+
+    enum class CardTag { CARD, MERCHANT, CORRUPTED }
+
     fun getTags(): List<CardTag> {
         return this.tags
     }
